get user remote_ip
* function to check user ip. * renamed the files.
